requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
google-cloud-storage==2.5.0
requests==2.28.1
python-dotenv==0.20.0
flask-restx==1.3.0
gunicorn 